@@ -20,8 +20,7 @@
     "jquery": "^3.1.1",
     "laravel-mix": "^1.0",
     "lodash": "^4.17.4",
-    "medium-editor-insert-plugin": "^2.3.2",
-    "vue": "^2.1.10"
+    "medium-editor-insert-plugin": "^2.3.2"
   },
   "dependencies": {
     "bootstrap-submenu": "^2.0.4",
@@ -32,8 +31,10 @@
     "jquery-ui": "^1.12.1",
     "mint-ui": "^2.2.7",
     "summernote": "^0.8.4",
+    "vue": "^2.5.4",
     "vue-chartjs": "^2.7.1",
     "vue-gesture": "^1.0.8",
+    "vue-router": "^3.0.1",
     "vue-touch": "^2.0.0-beta.4"
   }
 }
