@@ -1,4 +1,3 @@
-
 /**
  * First we will load all of this project's JavaScript dependencies which
  * includes Vue and other libraries. It is a great starting point when
@@ -7,9 +6,14 @@
 
 require('./bootstrap');
 
-window.Vue = require('vue');
+import Vue from 'vue';
+import VueRouter from 'vue-router';
 
-Vue.prototype.$http = window.axios;
+Vue.use(VueRouter);
+
+// window.Vue = require('vue');
+
+// Vue.prototype.$http = window.axios;
 
 /**
  * Next, we will create a fresh Vue application instance and attach it to
@@ -26,6 +30,25 @@ Vue.component('favorite', require('./components/Favorite.vue'));
 Vue.component('like', require('./components/Like.vue'));
 Vue.component('comment', require('./components/Comment.vue'));
 
-window.app = new Vue({
-    el: '#app'
-});
+
+Vue.component('subscriptions-left', require('./components/subscriptions/SubscriptionsLeft.vue'));
+
+const routes = [
+    { path: '/subscriptions_add', component: require('./components/subscriptions/SubscriptionsAdd.vue') },
+    { path: '/friend', component: require('./components/subscriptions/SubscriptionsFriend.vue') },
+    { path: '/category', component: require('./components/subscriptions/SubscriptionsCategory.vue') },
+    { path: '/collection', component: require('./components/subscriptions/SubscriptionsCollection.vue') },
+    { path: '/user', component: require('./components/subscriptions/SubscriptionsUser.vue') }
+];
+
+const router = new VueRouter({
+    routes
+})
+
+const app = new Vue({
+    router
+}).$mount('#app');
+
+// window.app = new Vue({
+//     el: '#app'
+// });
