@@ -1,46 +1,48 @@
 <footer id="tail">
-    <div class="container">
-        <div class="row">
-            <div class="col-xs-12">
-                <a href="#">
-                    关于爱你城
-                </a>
-                <em>
-                    ·
-                </em>
-                <a href="#">
-                    Copyright 2017 @ {{ env('APP_DOMAIN') }}
-                </a>
-                <em>
-                    ·
-                </em>
-                <a href="#">
-                    移动客户端
-                </a>
-                <em>
-                    ·
-                </em>
-                <a href="#">
-                    加入我们
-                </a>
-                <em>
-                    ·
-                </em>
-                <a href="#">
-                    爱你城投稿
-                </a>
-                <em>
-                    ·
-                </em>
-                <a href="#">
-                    帮助中心
-                </a>
-                <em>
-                    ·
-                </em>
-                <a href="#">
-                    友情链接
-                </a>
+    <div class="navbar-static-bottom">
+        <div class="container">
+            <div class="row">
+                <div class="col-xs-12">
+                    <a href="#">
+                        关于爱你城
+                    </a>
+                    <em>
+                        ·
+                    </em>
+                    <a href="#">
+                        Copyright 2017 @ {{ env('APP_DOMAIN') }}
+                    </a>
+                    <em>
+                        ·
+                    </em>
+                    <a href="#">
+                        移动客户端
+                    </a>
+                    <em>
+                        ·
+                    </em>
+                    <a href="#">
+                        加入我们
+                    </a>
+                    <em>
+                        ·
+                    </em>
+                    <a href="#">
+                        爱你城投稿
+                    </a>
+                    <em>
+                        ·
+                    </em>
+                    <a href="#">
+                        帮助中心
+                    </a>
+                    <em>
+                        ·
+                    </em>
+                    <a href="#">
+                        友情链接
+                    </a>
+                </div>
             </div>
         </div>
     </div>
