@@ -1,493 +1,439 @@
 @extends('v1.layouts.blank')
 
 @section('title')
-	空评 - 爱你城
+    空评 - 爱你城
 @stop
 @section('content')
-<div id="v1">
-    <div class="user">
-        <div class="container">
-            <div class="row">
-                <div class="essays col-xs-12 col-sm-8">
-                    <div class="main_top clearfix">
-                        <a class="avatar" href="/v1/category">
-                            <img src="//upload.jianshu.io/users/upload_avatars/4896574/81748b90-d20d-40fd-a659-127ece846249?imageMogr2/auto-orient/strip|imageView2/1/w/96/h/96"/>
-                        </a>
-                        <a class="botm follow" href="#">
-                            <i class="iconfont icon-jia">
-                            </i>
-                            <span>
-                                关注
-                            </span>
-                        </a>
-                        <a class="botm contribute" href="#">
-                            <span>
-                                发消息
-                            </span>
-                        </a>
-                        <a class="name" href="/v1/category">
+<div id="user">
+    <div class="container">
+        <div class="row">
+            <div class="essays col-xs-12 col-sm-8">
+                <div class="main_top clearfix">
+                    <a class="avatar" href="/v1/user">
+                        <img src="//upload.jianshu.io/users/upload_avatars/4896574/81748b90-d20d-40fd-a659-127ece846249?imageMogr2/auto-orient/strip|imageView2/1/w/96/h/96"/>
+                    </a>
+                    <a class="botm follow" href="#">
+                        <i class="iconfont icon-jia">
+                        </i>
+                        <span>
+                            关注
+                        </span>
+                    </a>
+                    <a class="botm contribute" href="#">
+                        <span>
+                            发消息
+                        </span>
+                    </a>
+                    <div class="title">
+                        <a class="name" href="/v1/user">
                             <span>
                                 空评
                             </span>
                             <i class="iconfont icon-nanhai">
                             </i>
                         </a>
-                        <p>
-                            收录了68923篇文章 · 1081552人关注
-                        </p>
                     </div>
-                    <ul class="trigger_menu">
-                        <li class="active">
-                            <a href="#">
+                    <div class="info">
+                        <ul>
+                            <li>
+                                <div class="meta_block">
+                                    <a href="#">
+                                        <p>
+                                            5
+                                        </p>
+                                        关注
+                                        <i class="iconfont icon-gengduo">
+                                        </i>
+                                    </a>
+                                </div>
+                            </li>
+                            <li>
+                                <div class="meta_block">
+                                    <a href="#">
+                                        <p>
+                                            6
+                                        </p>
+                                        粉丝
+                                        <i class="iconfont icon-gengduo">
+                                        </i>
+                                    </a>
+                                </div>
+                            </li>
+                            <li>
+                                <div class="meta_block">
+                                    <a href="#">
+                                        <p>
+                                            8
+                                        </p>
+                                        文章
+                                        <i class="iconfont icon-gengduo">
+                                        </i>
+                                    </a>
+                                </div>
+                            </li>
+                            <li>
+                                <div class="meta_block">
+                                    <p>
+                                        17299
+                                    </p>
+                                    <div>
+                                        字数
+                                    </div>
+                                </div>
+                            </li>
+                            <li>
+                                <div class="meta_block">
+                                    <p>
+                                        16
+                                    </p>
+                                    <div>
+                                        收获喜欢
+                                    </div>
+                                </div>
+                            </li>
+                        </ul>
+                    </div>
+                </div>
+                <div>
+                    <!-- Nav tabs -->
+                    <ul class="trigger_menu" role="tablist">
+                        <li class="active" role="presentation">
+                            <a aria-controls="wenzhang" data-toggle="tab" href="#wenzhang" role="tab">
                                 <i class="iconfont icon-dongtai1">
                                 </i>
                                 文章
                             </a>
                         </li>
-                        <li>
-                            <a href="#">
+                        <li role="presentation">
+                            <a aria-controls="dongtai" data-toggle="tab" href="#dongtai" role="tab">
                                 <i class="iconfont icon-dongtai">
                                 </i>
                                 动态
                             </a>
                         </li>
-                        <li>
-                            <a href="#">
+                        <li role="presentation">
+                            <a aria-controls="pinglun" data-toggle="tab" href="#pinglun" role="tab">
                                 <i class="iconfont icon-pinglun">
                                 </i>
                                 最新评论
                             </a>
                         </li>
-                        <li>
-                            <a href="#">
+                        <li role="presentation">
+                            <a aria-controls="huo" data-toggle="tab" href="#huo" role="tab">
                                 <i class="iconfont icon-huo">
                                 </i>
                                 热门
                             </a>
                         </li>
                     </ul>
-                    <ul class="list_container">
-                        <li>
-                            <div class="personal">
-                                <a class="portrait" href="/v1/user">
-                                    <img alt="头像" src="//upload.jianshu.io/users/upload_avatars/4896574/81748b90-d20d-40fd-a659-127ece846249?imageMogr2/auto-orient/strip|imageView2/1/w/96/h/96"/>
-                                </a>
-                                <div class="information">
-                                    <a href="/v1/user">
-                                        空评
-                                    </a>
-                                    <br/>
-                                    <small>
-                                        2天前
-                                    </small>
-                                </div>
-                            </div>
-                            <div class="substance have_img clearfix">
-                                <div class="conten col-xs-12 col-sm-8 col-md-9">
-                                    <a href="/v1/detail">
-                                        为什么说被马化腾点赞的《王者荣耀》已成为全球最赚钱的游戏？
-                                    </a>
-                                    <p class="article">
-                                        5月17日下午，腾讯控股公布了2017年第一季度财报。财报显示，腾讯一季度营收495.52亿元，同比增长55%；网络游戏收入增长34%至228.11亿元。其中，就智能手机游戏而言，腾讯实现129亿元收入，同比增长57%，此乃受现有及新的游戏如（《王者荣耀》、《穿越火线：枪战王者》及《龙之谷》）所推动。
-                                    </p>
-                                </div>
-                                <a class="wrap_img col-xs-10 col-sm-4 col-md-3 col-xs-offset-1 col-sm-offset-0" href="/v1/detail">
-                                    <img src="//upload-images.jianshu.io/upload_images/4896574-e0f2b6c752dab9c1?imageMogr2/auto-orient/strip%7CimageView2/2/w/1240"/>
-                                </a>
-                                <div class="statistics col-xs-12">
-                                    <a href="/v1/detail">
-                                        717阅读
-                                    </a>
-                                    <em>
-                                        ·
-                                    </em>
-                                    <a href="/v1/detail">
-                                        6评论
-                                    </a>
-                                    <em>
-                                        ·
-                                    </em>
-                                    <span>
-                                        13喜欢
-                                    </span>
-                                </div>
-                            </div>
-                        </li>
-                        <li>
-                            <div class="personal">
-                                <a class="portrait" href="/v1/user">
-                                    <img alt="头像" class="img-circle" src="//upload.jianshu.io/users/upload_avatars/1978017/8072afae5460?imageMogr2/auto-orient/strip|imageView2/1/w/96/h/96"/>
-                                </a>
-                                <div class="information">
-                                    <a href="/v1/user">
-                                        闫浩
-                                    </a>
-                                    <br/>
-                                    <small>
-                                        5天前
-                                    </small>
-                                </div>
-                            </div>
-                            <div class="substance have_img clearfix">
-                                <div class="conten col-xs-12 col-sm-8 col-md-9">
-                                    <a href="/v1/detail">
-                                        起风了，谁是下一个王者荣耀？
-                                    </a>
-                                    <p class="article">
-                                        下一个王者荣耀露出苗头很久了，今年夏天开始爆发。过去一年创投界有两个关注的焦点，一个是共享单车、共享充电宝带起来的线下流量入口，另一个是狼人杀和王者荣耀们的游戏社交化浪潮。这两个方向的创业者和巨头们，因缘际会地捕捉到了线上买量不再拥有性价比的历史脉搏，一方选择了开源，将物理投放的线下智能终端作为切入交易的新入口，另一方选择了节流，通过游戏本身的社交属性达成裂变式传播的效果，以较低的成本延长产品的生命周期以及活跃度。
-                                    </p>
-                                </div>
-                                <a class="wrap_img col-xs-10 col-sm-4 col-md-3 col-xs-offset-1 col-sm-offset-0" href="/v1/detail">
-                                    <img src="//upload-images.jianshu.io/upload_images/1978017-b9ea1315e388a415.jpg!1200?imageMogr2/auto-orient/strip%7CimageView2/2/w/1240"/>
-                                </a>
-                                <div class="statistics col-xs-12">
-                                    <a href="/v1/detail">
-                                        2325阅读
-                                    </a>
-                                    <em>
-                                        ·
-                                    </em>
-                                    <a href="/v1/detail">
-                                        5评论
-                                    </a>
-                                    <em>
-                                        ·
-                                    </em>
-                                    <span>
-                                        61喜欢
-                                    </span>
-                                    <em>
-                                        ·
-                                    </em>
-                                    <span>
-                                        2赞赏
-                                    </span>
-                                </div>
-                            </div>
-                        </li>
-                        <li>
-                            <div class="personal">
-                                <a class="portrait" href="/v1/user">
-                                    <img alt="头像" class="img-circle" src="//upload.jianshu.io/users/upload_avatars/1858256/e78657d29b8d.jpg?imageMogr2/auto-orient/strip|imageView2/1/w/96/h/96"/>
-                                </a>
-                                <div class="information">
-                                    <a href="/v1/user">
-                                        浮云笑此生
-                                    </a>
-                                    <br/>
-                                    <small>
-                                        3周前
-                                    </small>
-                                </div>
-                            </div>
-                            <div class="substance clearfix">
-                                <div class="conten col-xs-12">
-                                    <a href="/v1/detail">
-                                        相亲与自由恋爱的比较
-                                    </a>
-                                    <p class="article">
-                                        这里的“相亲”主要指的是一种东亚现象。相亲有传统的因素，但它没有遭遇传统普遍的危机，反而是历久弥新。相亲是通过第三方来寻找配偶，第三方可以是父母辈的亲人，可以是同辈的亲友，也可以是中介机构。不管是经由谁介绍，所有相亲的共同特点是：它是一种挑选。由此，相亲也是一种面试，只不过是平等的、双向的面试。这种挑选分为三步进行，第一步是“初试”，这往往是由第三方完成，相亲者并不参与但可能会给第三方提供一些条件；第二步是“复试”，进入复试的往往有多人，此时相亲者互相见面并初步沟通，这也是相亲之所谓相亲的主要方面；第三步是“终试”，往往只有一人进入这个阶段，此时他们往往已经成为实际上的情侣关系，这也是最后考验，直到步入婚姻殿堂为止。
-                                    </p>
-                                </div>
-                                {{--
-                                <a class="wrap_img col-xs-10 col-sm-4 col-md-3 col-xs-offset-1 col-sm-offset-0" href="/v1/detail">
-                                    <img src=""/>
-                                </a>
-                                --}}
-                                <div class="statistics col-xs-12">
-                                    <a href="/v1/detail">
-                                        1795阅读
-                                    </a>
-                                    <em>
-                                        ·
-                                    </em>
-                                    <a href="/v1/detail">
-                                        5评论
-                                    </a>
-                                    <em>
-                                        ·
-                                    </em>
-                                    <span>
-                                        35喜欢
-                                    </span>
-                                    <em>
-                                        ·
-                                    </em>
-                                    <span>
-                                        7赞赏
-                                    </span>
-                                </div>
-                            </div>
-                        </li>
-                        <li>
-                            <div class="personal">
-                                <a class="portrait" href="/v1/user">
-                                    <img alt="头像" class="img-circle" src="//upload.jianshu.io/users/upload_avatars/253140/b9adfdadef8a?imageMogr2/auto-orient/strip|imageView2/1/w/96/h/96"/>
-                                </a>
-                                <div class="information">
-                                    <a href="/v1/user">
-                                        迎刃
-                                    </a>
-                                    <br/>
-                                    <small>
-                                        1个月前
-                                    </small>
-                                </div>
-                            </div>
-                            <div class="substance have_img clearfix">
-                                <div class="conten col-xs-12 col-sm-8 col-md-9">
-                                    <a href="/v1/detail">
-                                        如何变得会与异性聊天？
-                                    </a>
-                                    <p class="article">
-                                        与人社交时如何很会聊天是个大家都非常感兴趣的话题，尤其是非常关注与异性聊天的问题。我综合了此前写过的若干篇文字，有了下面的关于聊天的方法论总结。从3个方面入手，让大家看完就能去实践。1，心态篇    2，原因篇    3，策略篇【心态篇】我发现如果没有一个良好的心态作为支撑和前提，做任何事情都会事倍功半。反之，心态好，自信心足，就会有强大的抗挫折承受力，甚至是遇强则强，越挫越勇。
-                                    </p>
-                                </div>
-                                <a class="wrap_img col-xs-10 col-sm-4 col-md-3 col-xs-offset-1 col-sm-offset-0" href="/v1/detail">
-                                    <img src="//upload-images.jianshu.io/upload_images/253140-5c2d51aa0d7c673d.jpg?imageMogr2/auto-orient/strip%7CimageView2/2/w/1240"/>
-                                </a>
-                                <div class="statistics col-xs-12">
-                                    <a href="/v1/detail">
-                                        317324阅读
-                                    </a>
-                                    <em>
-                                        ·
-                                    </em>
-                                    <a href="/v1/detail">
-                                        781评论
-                                    </a>
-                                    <em>
-                                        ·
-                                    </em>
-                                    <span>
-                                        11029喜欢
-                                    </span>
-                                    <em>
-                                        ·
-                                    </em>
-                                    <span>
-                                        12赞赏
-                                    </span>
-                                </div>
-                            </div>
-                        </li>
-                        <li>
-                            <div class="personal">
-                                <a class="portrait" href="/v1/user">
-                                    <img alt="头像" class="img-circle" src="//upload.jianshu.io/users/upload_avatars/6205434/831957b5-a672-4aa2-884b-cf971a8c34d5?imageMogr2/auto-orient/strip|imageView2/1/w/96/h/96"/>
-                                </a>
-                                <div class="information">
-                                    <a href="/v1/user">
-                                        网易王三三
-                                    </a>
-                                    <br/>
-                                    <small>
-                                        1个月 前
-                                    </small>
-                                </div>
-                            </div>
-                            <div class="substance have_img clearfix">
-                                <div class="conten col-xs-12 col-sm-8 col-md-9">
-                                    <a href="/v1/detail">
-                                        吃鸡和生活一样，九分天注定，一分靠打拼
-                                    </a>
-                                    <p class="article">
-                                        这几天，我发现大家的人生追求变了，从“明天不上班”变成了“今晚一定要吃到鸡”。“吃鸡”梗来源于最近火爆全球的一款射击类生存游戏《绝地求生：大逃杀》。主要是讲几十上百个猛男被投放到一个荒岛，为了最终目标——吃鸡而互相厮杀，干掉所有对手活到最后的故事。当然，它还有其他一些流传更广的名字，比如《伏地求生之老阴X大乱斗》《一千零一种死法》《幻影坦克大战千年老苟》等。
-                                    </p>
-                                </div>
-                                <a class="wrap_img col-xs-10 col-sm-4 col-md-3 col-xs-offset-1 col-sm-offset-0" href="/v1/detail">
-                                    <img src="//upload-images.jianshu.io/upload_images/6205434-62e64dc2d81e3625.jpeg?imageMogr2/auto-orient/strip%7CimageView2/2/w/1240"/>
-                                </a>
-                                <div class="statistics col-xs-12">
-                                    <a href="/v1/detail">
-                                        834阅读
-                                    </a>
-                                    <em>
-                                        ·
-                                    </em>
-                                    <a href="/v1/detail">
-                                        6评论
-                                    </a>
-                                    <em>
-                                        ·
-                                    </em>
-                                    <span>
-                                        27喜欢
-                                    </span>
-                                </div>
-                            </div>
-                        </li>
-                        <li>
-                            <div class="personal">
-                                <a class="portrait" href="/v1/user">
-                                    <img alt="头像" class="img-circle" src="//upload.jianshu.io/users/upload_avatars/6486956/8b5f880e-f7ba-446b-8b59-1532e087fe3e.jpeg?imageMogr2/auto-orient/strip|imageView2/1/w/96/h/96"/>
-                                </a>
-                                <div class="information">
-                                    <a href="/v1/user">
-                                        智_先生
-                                    </a>
-                                    <br/>
-                                    <small>
-                                        2个月 前
-                                    </small>
-                                </div>
-                            </div>
-                            <div class="substance have_img clearfix">
-                                <div class="conten col-xs-12 col-sm-8 col-md-9">
-                                    <a href="/v1/detail">
-                                        我们该感谢王者荣耀耗尽了80%的人上升的空间
-                                    </a>
-                                    <p class="article">
-                                        在当下，以《王者荣耀》、《阴阳师》为首的一批手游正逐渐霸占人们的手机屏幕，以此而衍生的电竞、直播以及代练行业风光无限，潜力巨大。最近，为保障未成年人健康成长，腾讯经过将近一个月的调试和内测后，将于7月4日以《王者荣耀》为试点，率先推出健康游戏防沉迷系统的“三板斧”，对未成年人的游戏时间作出限制：12岁以下每天限玩一小时。（小学生听了绝不会流泪，并笑着拿出奶奶的身份证。）当前国内还没有移动游戏防沉迷的明确规定，《王者荣耀》在顶着“亡者毒药”的社会舆论中做出了实际行动，这是一种自保行为，也可以说是社会责任感的体现。但事实上，《王者荣耀》的最大玩家群体当属大学生和中年人群体，第三方数据分析机构talkingData统计的数据显示，王者荣耀的用户群中，大学生占21.8%，中小学生占2.7%；上班族占68.7%。
-                                    </p>
-                                </div>
-                                <a class="wrap_img col-xs-10 col-sm-4 col-md-3 col-xs-offset-1 col-sm-offset-0" href="/v1/detail">
-                                    <img src="//upload-images.jianshu.io/upload_images/6486956-1ce7c76170f7b916.png?imageMogr2/auto-orient/strip%7CimageView2/2/w/1240"/>
-                                </a>
-                                <div class="statistics col-xs-12">
-                                    <a href="/v1/detail">
-                                        113787阅读
-                                    </a>
-                                    <em>
-                                        ·
-                                    </em>
-                                    <a href="/v1/detail">
-                                        1096评论
-                                    </a>
-                                    <em>
-                                        ·
-                                    </em>
-                                    <span>
-                                        3222喜欢
-                                    </span>
-                                    <em>
-                                        ·
-                                    </em>
-                                    <span>
-                                        49赞赏
-                                    </span>
-                                </div>
-                            </div>
-                        </li>
-                        <li>
-                            <div class="personal">
-                                <a class="portrait" href="/v1/user">
-                                    <img alt="头像" class="img-circle" src="//upload.jianshu.io/users/upload_avatars/1408329/1ed19a008804?imageMogr2/auto-orient/strip|imageView2/1/w/96/h/96"/>
-                                </a>
-                                <div class="information">
-                                    <a href="/v1/user">
-                                        冷清持
-                                    </a>
-                                    <br/>
-                                    <small>
-                                        2个月 前
-                                    </small>
-                                </div>
-                            </div>
-                            <div class="substance have_img clearfix">
-                                <div class="conten col-xs-12 col-sm-8 col-md-9">
-                                    <a href="/v1/detail">
-                                        和男朋友一起玩游戏是一种怎样的感觉
-                                    </a>
-                                    <p class="article">
-                                        搬新家了，别人都是在考虑装修，买家具，整理房间，而作为一个殿堂级游戏迷，我和男朋友却一致把最大重心放在了买电脑上。区别是我以前喜欢玩单机，男朋友喜欢玩网游。﻿结果当然是买了一模一样的两台高配。﻿某天我们躺在沙发上，他突然说，哎宝宝，不如我带你玩个游戏吧。﻿我弯眉:“好啊好啊是你拖地我加油的游戏吗。”﻿他无视:“一起玩游戏很棒的呀，可以建个小号慢慢升级啊，一起打副本啊，一起钓鱼啊，看风景啦…是不是感觉超棒？我朋友也在玩我们可以组个固定队！”﻿
-                                    </p>
-                                </div>
-                                <a class="wrap_img col-xs-10 col-sm-4 col-md-3 col-xs-offset-1 col-sm-offset-0" href="/v1/detail">
-                                    <img src="http://upload-images.jianshu.io/upload_images/1408329-d38d06f1f04211fd.jpg?imageMogr2/auto-orient/strip%7CimageView2/2/w/1080/q/100"/>
-                                </a>
-                                <div class="statistics col-xs-12">
-                                    <a href="/v1/detail">
-                                        11898阅读
-                                    </a>
-                                    <em>
-                                        ·
-                                    </em>
-                                    <a href="/v1/detail">
-                                        299评论
-                                    </a>
-                                    <em>
-                                        ·
-                                    </em>
-                                    <span>
-                                        303喜欢
-                                    </span>
-                                    <em>
-                                        ·
-                                    </em>
-                                    <span>
-                                        5赞赏
-                                    </span>
-                                </div>
-                            </div>
-                        </li>
-                    </ul>
-                    <a class="load_more" href="#">
-                        阅读更多
+                    <!-- Tab panes -->
+                    <div class="tab-content">
+                        <div class="tab-pane active" id="wenzhang" role="tabpanel">
+                            @include('v1.parts.writing')
+                        </div>
+                        <div class="tab-pane" id="dongtai" role="tabpanel">
+                            <ul class="list_container">
+                                <li>
+                                    <div class="personal">
+                                        <a class="avatar" href="/v1/user">
+                                            <img alt="头像" src="//upload.jianshu.io/users/upload_avatars/4896574/81748b90-d20d-40fd-a659-127ece846249?imageMogr2/auto-orient/strip|imageView2/1/w/96/h/96"/>
+                                        </a>
+                                        <div class="information">
+                                            <a href="/v1/user">
+                                                空评
+                                            </a>
+                                            <small>
+                                                发表了文章 · 06.15.12:12
+                                            </small>
+                                        </div>
+                                    </div>
+                                    <div class="substance have_img clearfix">
+                                        <div class="conten col-xs-12 col-sm-8 col-md-9">
+                                            <a href="/v1/detail">
+                                                魔兽世界7.2.5全新版本资料片：新橙戒+新黑庙+乐队活动
+                                            </a>
+                                            <p class="article">
+                                                《魔兽世界》7.2.5版本今日在国服上线。虽然是小版本更新，但改动内容还是十分丰富的。比如新的橙装、大秘境调整、黑庙加入时空漫游、克罗米战役等等。以下是17173为各位整理的《魔兽世界》7.2.5版本今日在国服上线。虽然是小版本更新，但改动内容还是十分丰富的。比如新的橙装、大秘境调整、黑庙加入时空漫游、克罗米战役等等。以下是17173为各位整理的
+                                            </p>
+                                        </div>
+                                        <a class="wrap_img col-xs-10 col-sm-4 col-md-3 col-xs-offset-1 col-sm-offset-0" href="/v1/detail">
+                                            <img src="//upload-images.jianshu.io/upload_images/4896574-5091b316532c0d99?imageMogr2/auto-orient/strip|imageView2/1/w/150/h/120"/>
+                                        </a>
+                                        <div class="statistics col-xs-12">
+                                            <a href="/v1/detail">
+                                                <i class="iconfont icon-yanjing">
+                                                </i>
+                                                717
+                                            </a>
+                                            <a href="/v1/detail">
+                                                <i class="iconfont icon-weibiaoti5">
+                                                </i>
+                                                6
+                                            </a>
+                                            <span>
+                                                <i class="iconfont icon-xihuan1">
+                                                </i>
+                                                13
+                                            </span>
+                                        </div>
+                                    </div>
+                                </li>
+                                <li>
+                                    <div class="personal">
+                                        <a class="avatar" href="/v1/user">
+                                            <img alt="头像" src="//upload.jianshu.io/users/upload_avatars/4896574/81748b90-d20d-40fd-a659-127ece846249?imageMogr2/auto-orient/strip|imageView2/1/w/96/h/96"/>
+                                        </a>
+                                        <div class="information">
+                                            <a href="/v1/user">
+                                                空评
+                                            </a>
+                                            <small>
+                                                关注了作者 · 06.14.23:55
+                                            </small>
+                                        </div>
+                                    </div>
+                                    <div class="follow_detail">
+                                        <div class="info">
+                                            <a class="avatar" href="#">
+                                                <img src="//upload.jianshu.io/users/upload_avatars/1159369/799d8fa18062?imageMogr2/auto-orient/strip|imageView2/1/w/180/h/180"/>
+                                            </a>
+                                            <a class="follow" href="#">
+                                                <i class="iconfont icon-jia">
+                                                </i>
+                                                <span>
+                                                    关注
+                                                </span>
+                                            </a>
+                                            <a class="title" href="#">
+                                                Dreamover1010
+                                            </a>
+                                            <p>
+                                                写了 16299 字，被 90 人关注，获得了 140 个喜欢
+                                            </p>
+                                        </div>
+                                        <div class="signature">
+                                            在广告行业摸爬滚打几年的设计师逃出升天
+                                        </div>
+                                    </div>
+                                </li>
+                                <li>
+                                    <div class="personal">
+                                        <a class="avatar" href="/v1/user">
+                                            <img alt="头像" src="//upload.jianshu.io/users/upload_avatars/4896574/81748b90-d20d-40fd-a659-127ece846249?imageMogr2/auto-orient/strip|imageView2/1/w/96/h/96"/>
+                                        </a>
+                                        <div class="information">
+                                            <a href="/v1/user">
+                                                空评
+                                            </a>
+                                            <small>
+                                                发表了评论 · 10.30.13:26
+                                            </small>
+                                        </div>
+                                    </div>
+                                    <div class="comment">
+                                        <p>
+                                            <a href="#">
+                                                @夜_2d81
+                                            </a>
+                                            丧尸从0到1%，这个过程，军队就开始介入了，基本不可能达到99%
+                                        </p>
+                                        <blockquote>
+                                            <a class="title" href="#">
+                                                末日来临，最正确的丧尸自救指南
+                                            </a>
+                                            <p class="abstract">
+                                                这是智先生的第8篇原创文章 当末日来临 你准备好了吗？ 最近我看完了马克斯·布鲁克斯写的《僵尸生存指南》，书本教会了人们如何在丧尸横行的世界中生存，并提供了从选择武器到制定逃...
+                                            </p>
+                                            <div class="meta">
+                                                <div class="origin_author">
+                                                    <a href="#">
+                                                        空评
+                                                    </a>
+                                                </div>
+                                                <a href="/v1/detail">
+                                                    <i class="iconfont icon-yanjing">
+                                                    </i>
+                                                    1649
+                                                </a>
+                                                <a href="/v1/detail">
+                                                    <i class="iconfont icon-weibiaoti5">
+                                                    </i>
+                                                    22
+                                                </a>
+                                                <span>
+                                                    <i class="iconfont icon-xihuan1">
+                                                    </i>
+                                                    72
+                                                </span>
+                                            </div>
+                                        </blockquote>
+                                    </div>
+                                </li>
+                                <li>
+                                    <div class="personal">
+                                        <a class="avatar" href="/v1/user">
+                                            <img alt="头像" src="//upload.jianshu.io/users/upload_avatars/4896574/81748b90-d20d-40fd-a659-127ece846249?imageMogr2/auto-orient/strip|imageView2/1/w/96/h/96"/>
+                                        </a>
+                                        <div class="information">
+                                            <a href="/v1/user">
+                                                空评
+                                            </a>
+                                            <em>
+                                                ·
+                                            </em>
+                                            <small>
+                                                赞了评论 · 10.30.13:26
+                                            </small>
+                                        </div>
+                                    </div>
+                                    <div class="comment">
+                                        <p>
+                                            为这篇文章注册简书，我也挺难受大学寝室的环境的，本身自己挺孤僻，习惯一个人待着，最不能忍的是大二的洗澡是十几个开放式没有门的洗澡间···庆幸室友的性格都挺好，现在毕业两年还在联系，今天还和其中一个说想给她设计婚礼请帖的图案的，我玩的很好的高中同学大学5年的室友则是比较针对她，好像是因为拿奖学金之类的事吧，然后处处挤兑，她研究生的室友则是家教啥的都很好，她现在可开心了 我也很替她开心，她读研和我工作在一个城市
+                                        </p>
+                                        <blockquote>
+                                            <div class="meta">
+                                                <div class="origin_author">
+                                                    <a href="#">
+                                                        郭璐Alu
+                                                    </a>
+                                                    <span>
+                                                        评论自
+                                                        <a href="#">
+                                                            集体宿舍，将3000万大学生拉入深渊
+                                                        </a>
+                                                    </span>
+                                                </div>
+                                            </div>
+                                        </blockquote>
+                                    </div>
+                                </li>
+                            </ul>
+                        </div>
+                        <div class="tab-pane" id="pinglun" role="tabpanel">
+                            @include('v1.parts.writing')
+                        </div>
+                        <div class="tab-pane" id="huo" role="tabpanel">
+                            @include('v1.parts.writing')
+                        </div>
+                    </div>
+                </div>
+            </div>
+            <div class="aside col-sm-4 col-lg-3 col-lg-offset-1">
+                <ul class="user_dynamic">
+                    <li>
+                        <a href="#">
+                            <img src="//cdn2.jianshu.io/assets/badges/signed-9702260821906f0d953eab67a29f8e7a2d2e3d20960576347591283a3fbfd867.png"/>
+                            爱你城签约作者
+                        </a>
+                    </li>
+                    <li>
+                        <a href="#">
+                            <img src="//cdn2.jianshu.io/assets/badges/excellent-34cdda4de26ee9910f190239ffecf886c24045136468fa26f041b0156a143cd9.png"/>
+                            空评《好中文的样子》
+                        </a>
+                    </li>
+                </ul>
+                <p class="title">
+                    个人介绍
+                </p>
+                <div class="description">
+                    <p>
+                        十分精力：四分读书，三分写书，三分教书。《正版语文》作者，21世纪最牛写作书 《风格感觉》（The Sense of Style）中文译者之一，“好中文的样子”创办人和主讲人。
+                    </p>
+                    <a href="javascript:">
+                        <i class="iconfont icon-xinlangweibo1">
+                        </i>
+                    </a>
+                    <a href="javascript:">
+                        <i class="iconfont icon-weixin">
+                        </i>
                     </a>
                 </div>
-                <div class="aside col-sm-4 col-lg-3 col-lg-offset-1">
-                    <p class="title">
-                        个人介绍
-                    </p>
-                    <div class="description">
-                        <p>
-                            暂无个人介绍
-                        </p>
-                        {{--
-                        <a href="javascript:">
-                            <i class="iconfont icon-xinlangweibo1">
+                <ul class="user_dynamic">
+                    <li>
+                        <a href="#">
+                            <i class="iconfont icon-combinedshape">
                             </i>
+                            <span>
+                                他关注的专题/文集
+                            </span>
                         </a>
-                        --}}
-                    </div>
-                    <ul class="user_dynamic">
+                    </li>
+                    <li>
+                        <a href="#">
+                            <i class="iconfont icon-xihuan">
+                            </i>
+                            <span>
+                                他喜欢的文章
+                            </span>
+                        </a>
+                    </li>
+                </ul>
+                <div>
+                    <p class="title">
+                        他管理的专题
+                    </p>
+                    <ul class="list">
                         <li>
                             <a href="#">
-                                <i class="iconfont icon-zhuanti">
-                                </i>
+                                <img src="//upload.jianshu.io/collections/images/367981/android.graphics.Bitmap_2bebf481.jpeg?imageMogr2/auto-orient/strip|imageView2/1/w/96/h/96"/>
                                 <span>
-                                    他关注的专题/文集
+                                    文章
                                 </span>
                             </a>
                         </li>
                         <li>
                             <a href="#">
-                                <i class="iconfont icon-xihuan">
-                                </i>
+                                <img src="//upload.jianshu.io/collections/images/319558/1.jpg?imageMogr2/auto-orient/strip|imageView2/1/w/96/h/96"/>
                                 <span>
-                                    他喜欢的文章
+                                    好中文的样子
                                 </span>
                             </a>
                         </li>
                     </ul>
-                    <div>
-                        <p class="title">
-                            他的文集
-                        </p>
-                        <ul class="list">
-                            <li>
-                                <a href="#">
-                                    <i class="iconfont icon-wenji">
-                                    </i>
-                                    <span>
-                                        日记本
-                                    </span>
-                                </a>
-                            </li>
-                        </ul>
-                    </div>
-                    <div class="user_action">
-                        <a href="#">
-                            加入黑名单
-                        </a>
-                        <em>
-                            ·
-                        </em>
-                        <a href="#">
-                            举报用户
-                        </a>
-                    </div>
+                    <p class="title">
+                        他的文集
+                    </p>
+                    <ul class="list">
+                        <li>
+                            <a href="#">
+                                <i class="iconfont icon-wenji">
+                                </i>
+                                <span>
+                                    日记本
+                                </span>
+                            </a>
+                        </li>
+                        <li>
+                            <a href="#">
+                                <i class="iconfont icon-wenji">
+                                </i>
+                                <span>
+                                    好中文的样子
+                                </span>
+                            </a>
+                        </li>
+                        <li>
+                            <a href="#">
+                                <i class="iconfont icon-wenji">
+                                </i>
+                                <span>
+                                    个人试笔
+                                </span>
+                            </a>
+                        </li>
+                    </ul>
+                </div>
+                <div class="user_action">
+                    <a href="#">
+                        加入黑名单
+                    </a>
+                    <em>
+                        ·
+                    </em>
+                    <a href="#">
+                        举报用户
+                    </a>
                 </div>
             </div>
         </div>
